rename: 경매 목록 조회 관련 테스트 패키지 분리
