+++ backend/ddang/src/test/java/com/ddang/ddang/auction/infrastructure/persistence/fixture/list/AuctionForListSortByAuctioneerCountFixture.java
@@ -1,6 +1,7 @@
-package com.ddang.ddang.auction.infrastructure.persistence.fixture;
+package com.ddang.ddang.auction.infrastructure.persistence.fixture.list;
 
 import com.ddang.ddang.auction.domain.Auction;
+import com.ddang.ddang.auction.infrastructure.persistence.fixture.QuerydslAuctionRepositoryForListFixture;
 import com.ddang.ddang.auction.presentation.dto.request.ReadAuctionSearchCondition;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.data.domain.Sort;
