+++ backend/ddang/src/test/java/com/ddang/ddang/auction/infrastructure/persistence/fixture/list/AuctionForListSortByReliabilityFixture.java
@@ -1,16 +1,17 @@
-package com.ddang.ddang.auction.infrastructure.persistence.fixture;
+package com.ddang.ddang.auction.infrastructure.persistence.fixture.list;
 
 import com.ddang.ddang.auction.domain.Auction;
+import com.ddang.ddang.auction.infrastructure.persistence.fixture.QuerydslAuctionRepositoryForListFixture;
 import com.ddang.ddang.auction.presentation.dto.request.ReadAuctionSearchCondition;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.domain.Sort.Order;
 
 @SuppressWarnings("NonAsciiCharacters")
-public class AuctionForListSearchByTitleAndSortByReliabilityFixture extends QuerydslAuctionRepositoryForListFixture {
+public class AuctionForListSortByReliabilityFixture extends QuerydslAuctionRepositoryForListFixture {
 
     protected Sort 신뢰도순_정렬 = Sort.by(Order.asc("reliability"));
-    protected ReadAuctionSearchCondition 검색어_맥북 = new ReadAuctionSearchCondition("맥북");
+    protected ReadAuctionSearchCondition 검색어_없음 = new ReadAuctionSearchCondition(null);
 
     protected Auction 첫번째_페이지_인덱스_0_신뢰도_5_0_2일_후_마감_id_4;
     protected Auction 첫번째_페이지_인덱스_1_신뢰도_3_5_4일_후_마감_id_14;
@@ -20,12 +21,14 @@ public class AuctionForListSearchByTitleAndSortByReliabilityFixture extends Quer
     protected Auction 두번째_페이지_인덱스_2_신뢰도_2_1_4일_후_마감_id_11;
     protected Auction 세번째_페이지_인덱스_0_신뢰도_2_1_3일_후_마감_id_3;
     protected Auction 세번째_페이지_인덱스_1_신뢰도_2_1_4일_후_마감_id_2;
-    protected Auction 세번째_페이지_인덱스_2_신뢰도_4_7_4일_전_마감_id_7;
-    protected Auction 네번째_페이지_인덱스_0_신뢰도_4_7_5일_전_마감_id_1;
-    protected Auction 네번째_페이지_인덱스_1_신뢰도_3_5_4일_전_마감_id_16;
-    protected Auction 네번째_페이지_인덱스_2_신뢰도_3_5_4일_전_마감_id_8;
-    protected Auction 다섯번째_페이지_인덱스_0_신뢰도_2_1_4일_전_마감_id_13;
-    protected Auction 다섯번째_페이지_인덱스_1_신뢰도_2_1_4일_전_마감_id_9;
+    protected Auction 세번째_페이지_인덱스_2_신뢰도_1_5_1일_후_마감_id_5;
+    protected Auction 네번째_페이지_인덱스_0_신뢰도_4_7_4일_전_마감_id_7;
+    protected Auction 네번째_페이지_인덱스_1_신뢰도_4_7_5일_전_마감_id_1;
+    protected Auction 네번째_페이지_인덱스_2_신뢰도_3_5_4일_전_마감_id_16;
+    protected Auction 다섯번째_페이지_인덱스_0_신뢰도_3_5_4일_전_마감_id_8;
+    protected Auction 다섯번째_페이지_인덱스_1_신뢰도_2_1_4일_전_마감_id_13;
+    protected Auction 다섯번째_페이지_인덱스_2_신뢰도_2_1_4일_전_마감_id_9;
+    protected Auction 여섯번째_페이지_인덱스_0_신뢰도_0_3_2일_전_마감_id_6;
 
     @BeforeEach
     void fixtureSetUp() {
@@ -37,11 +40,13 @@ public class AuctionForListSearchByTitleAndSortByReliabilityFixture extends Quer
         두번째_페이지_인덱스_2_신뢰도_2_1_4일_후_마감_id_11 = 경매11;
         세번째_페이지_인덱스_0_신뢰도_2_1_3일_후_마감_id_3 = 경매3;
         세번째_페이지_인덱스_1_신뢰도_2_1_4일_후_마감_id_2 = 경매2;
-        세번째_페이지_인덱스_2_신뢰도_4_7_4일_전_마감_id_7 = 경매7;
-        네번째_페이지_인덱스_0_신뢰도_4_7_5일_전_마감_id_1 = 경매1;
-        네번째_페이지_인덱스_1_신뢰도_3_5_4일_전_마감_id_16 = 경매16;
-        네번째_페이지_인덱스_2_신뢰도_3_5_4일_전_마감_id_8 = 경매8;
-        다섯번째_페이지_인덱스_0_신뢰도_2_1_4일_전_마감_id_13 = 경매13;
-        다섯번째_페이지_인덱스_1_신뢰도_2_1_4일_전_마감_id_9 = 경매9;
+        세번째_페이지_인덱스_2_신뢰도_1_5_1일_후_마감_id_5 = 경매5;
+        네번째_페이지_인덱스_0_신뢰도_4_7_4일_전_마감_id_7 = 경매7;
+        네번째_페이지_인덱스_1_신뢰도_4_7_5일_전_마감_id_1 = 경매1;
+        네번째_페이지_인덱스_2_신뢰도_3_5_4일_전_마감_id_16 = 경매16;
+        다섯번째_페이지_인덱스_0_신뢰도_3_5_4일_전_마감_id_8 = 경매8;
+        다섯번째_페이지_인덱스_1_신뢰도_2_1_4일_전_마감_id_13 = 경매13;
+        다섯번째_페이지_인덱스_2_신뢰도_2_1_4일_전_마감_id_9 = 경매9;
+        여섯번째_페이지_인덱스_0_신뢰도_0_3_2일_전_마감_id_6 = 경매6;
     }
 }
