+++ backend/ddang/src/test/java/com/ddang/ddang/auction/infrastructure/persistence/list/AuctionForListByUserIdTest.java
@@ -1,7 +1,8 @@
-package com.ddang.ddang.auction.infrastructure.persistence;
+package com.ddang.ddang.auction.infrastructure.persistence.list;
 
 import com.ddang.ddang.auction.domain.Auction;
-import com.ddang.ddang.auction.infrastructure.persistence.fixture.AuctionForListByUserIdFixture;
+import com.ddang.ddang.auction.infrastructure.persistence.QuerydslAuctionRepository;
+import com.ddang.ddang.auction.infrastructure.persistence.fixture.list.AuctionForListByUserIdFixture;
 import com.ddang.ddang.configuration.JpaConfiguration;
 import com.ddang.ddang.configuration.QuerydslConfiguration;
 import com.querydsl.jpa.impl.JPAQueryFactory;
