+++ backend/ddang/src/test/java/com/ddang/ddang/auction/infrastructure/persistence/list/AuctionForListSearchByTitleAndSortByIdTest.java
@@ -1,7 +1,8 @@
-package com.ddang.ddang.auction.infrastructure.persistence;
+package com.ddang.ddang.auction.infrastructure.persistence.list;
 
 import com.ddang.ddang.auction.domain.Auction;
-import com.ddang.ddang.auction.infrastructure.persistence.fixture.AuctionForListSearchByTitleAndSortByClosingTimeFixture;
+import com.ddang.ddang.auction.infrastructure.persistence.QuerydslAuctionRepository;
+import com.ddang.ddang.auction.infrastructure.persistence.fixture.list.AuctionForListSearchByTitleAndSortByIdFixture;
 import com.ddang.ddang.configuration.JpaConfiguration;
 import com.ddang.ddang.configuration.QuerydslConfiguration;
 import com.querydsl.jpa.impl.JPAQueryFactory;
@@ -20,7 +21,7 @@ import org.springframework.data.domain.Slice;
 @Import({JpaConfiguration.class, QuerydslConfiguration.class})
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
 @SuppressWarnings("NonAsciiCharacters")
-class AuctionForListSearchByTitleAndSortByClosingTimeTest extends AuctionForListSearchByTitleAndSortByClosingTimeFixture {
+class AuctionForListSearchByTitleAndSortByIdTest extends AuctionForListSearchByTitleAndSortByIdFixture {
 
     QuerydslAuctionRepository querydslAuctionRepository;
 
@@ -34,15 +35,15 @@ class AuctionForListSearchByTitleAndSortByClosingTimeTest extends AuctionForList
         // when
         final Slice<Auction> actual = querydslAuctionRepository.findAuctionsAllByCondition(
                 검색어_맥북,
-                PageRequest.of(0, 페이지_크기_3, 마감_임박순_정렬)
+                PageRequest.of(0, 페이지_크기_3, id순_정렬)
         );
 
         // then
         SoftAssertions.assertSoftly(softAssertions -> {
             softAssertions.assertThat(actual).hasSize(3);
-            softAssertions.assertThat(actual.getContent().get(0)).isEqualTo(첫번째_페이지_인덱스_0_2일_후_마감_id_4);
-            softAssertions.assertThat(actual.getContent().get(1)).isEqualTo(첫번째_페이지_인덱스_1_3일_후_마감_id_3);
-            softAssertions.assertThat(actual.getContent().get(2)).isEqualTo(첫번째_페이지_인덱스_2_4일_후_마감_id_15);
+            softAssertions.assertThat(actual.getContent().get(0)).isEqualTo(첫번째_페이지_인덱스_0_4일_후_마감_id_15);
+            softAssertions.assertThat(actual.getContent().get(1)).isEqualTo(첫번째_페이지_인덱스_1_4일_후_마감_id_14);
+            softAssertions.assertThat(actual.getContent().get(2)).isEqualTo(첫번째_페이지_인덱스_2_4일_후_마감_id_12);
             softAssertions.assertThat(actual.hasNext()).isTrue();
         });
     }
@@ -52,15 +53,15 @@ class AuctionForListSearchByTitleAndSortByClosingTimeTest extends AuctionForList
         // when
         final Slice<Auction> actual = querydslAuctionRepository.findAuctionsAllByCondition(
                 검색어_맥북,
-                PageRequest.of(1, 페이지_크기_3, 마감_임박순_정렬)
+                PageRequest.of(1, 페이지_크기_3, id순_정렬)
         );
 
         // then
         SoftAssertions.assertSoftly(softAssertions -> {
             softAssertions.assertThat(actual).hasSize(3);
-            softAssertions.assertThat(actual.getContent().get(0)).isEqualTo(두번째_페이지_인덱스_0_4일_후_마감_id_14);
-            softAssertions.assertThat(actual.getContent().get(1)).isEqualTo(두번째_페이지_인덱스_1_4일_후_마감_id_12);
-            softAssertions.assertThat(actual.getContent().get(2)).isEqualTo(두번째_페이지_인덱스_2_4일_후_마감_id_11);
+            softAssertions.assertThat(actual.getContent().get(0)).isEqualTo(두번째_페이지_인덱스_0_4일_후_마감_id_11);
+            softAssertions.assertThat(actual.getContent().get(1)).isEqualTo(두번째_페이지_인덱스_1_4일_후_마감_id_10);
+            softAssertions.assertThat(actual.getContent().get(2)).isEqualTo(두번째_페이지_인덱스_2_2일_후_마감_id_4);
             softAssertions.assertThat(actual.hasNext()).isTrue();
         });
     }
@@ -70,15 +71,15 @@ class AuctionForListSearchByTitleAndSortByClosingTimeTest extends AuctionForList
         // when
         final Slice<Auction> actual = querydslAuctionRepository.findAuctionsAllByCondition(
                 검색어_맥북,
-                PageRequest.of(2, 페이지_크기_3, 마감_임박순_정렬)
+                PageRequest.of(2, 페이지_크기_3, id순_정렬)
         );
 
         // then
         SoftAssertions.assertSoftly(softAssertions -> {
             softAssertions.assertThat(actual).hasSize(3);
-            softAssertions.assertThat(actual.getContent().get(0)).isEqualTo(세번째_페이지_인덱스_0_4일_후_마감_id_10);
+            softAssertions.assertThat(actual.getContent().get(0)).isEqualTo(세번째_페이지_인덱스_0_3일_후_마감_id_3);
             softAssertions.assertThat(actual.getContent().get(1)).isEqualTo(세번째_페이지_인덱스_1_4일_후_마감_id_2);
-            softAssertions.assertThat(actual.getContent().get(2)).isEqualTo(세번째_페이지_인덱스_2_5일_전_마감_id_1);
+            softAssertions.assertThat(actual.getContent().get(2)).isEqualTo(세번째_페이지_인덱스_2_4일_전_마감_id_16);
             softAssertions.assertThat(actual.hasNext()).isTrue();
         });
     }
@@ -88,15 +89,15 @@ class AuctionForListSearchByTitleAndSortByClosingTimeTest extends AuctionForList
         // when
         final Slice<Auction> actual = querydslAuctionRepository.findAuctionsAllByCondition(
                 검색어_맥북,
-                PageRequest.of(3, 페이지_크기_3, 마감_임박순_정렬)
+                PageRequest.of(3, 페이지_크기_3, id순_정렬)
         );
 
         // then
         SoftAssertions.assertSoftly(softAssertions -> {
             softAssertions.assertThat(actual).hasSize(3);
-            softAssertions.assertThat(actual.getContent().get(0)).isEqualTo(네번째_페이지_인덱스_0_4일_전_마감_id_16);
-            softAssertions.assertThat(actual.getContent().get(1)).isEqualTo(네번째_페이지_인덱스_1_4일_전_마감_id_13);
-            softAssertions.assertThat(actual.getContent().get(2)).isEqualTo(네번째_페이지_인덱스_2_4일_전_마감_id_9);
+            softAssertions.assertThat(actual.getContent().get(0)).isEqualTo(네번째_페이지_인덱스_0_4일_전_마감_id_13);
+            softAssertions.assertThat(actual.getContent().get(1)).isEqualTo(네번째_페이지_인덱스_1_4일_전_마감_id_9);
+            softAssertions.assertThat(actual.getContent().get(2)).isEqualTo(네번째_페이지_인덱스_2_4일_전_마감_id_8);
             softAssertions.assertThat(actual.hasNext()).isTrue();
         });
     }
@@ -106,13 +107,13 @@ class AuctionForListSearchByTitleAndSortByClosingTimeTest extends AuctionForList
         // when
         final Slice<Auction> actual = querydslAuctionRepository.findAuctionsAllByCondition(
                 검색어_맥북,
-                PageRequest.of(4, 페이지_크기_3, 마감_임박순_정렬)
+                PageRequest.of(4, 페이지_크기_3, id순_정렬)
         );
 
         SoftAssertions.assertSoftly(softAssertions -> {
             softAssertions.assertThat(actual).hasSize(2);
-            softAssertions.assertThat(actual.getContent().get(0)).isEqualTo(다섯번째_페이지_인덱스_0_4일_전_마감_id_8);
-            softAssertions.assertThat(actual.getContent().get(1)).isEqualTo(다섯번째_페이지_인덱스_1_4일_전_마감_id_7);
+            softAssertions.assertThat(actual.getContent().get(0)).isEqualTo(다섯번째_페이지_인덱스_0_4일_전_마감_id_7);
+            softAssertions.assertThat(actual.getContent().get(1)).isEqualTo(다섯번째_페이지_인덱스_1_5일_전_마감_id_1);
             softAssertions.assertThat(actual.hasNext()).isFalse();
         });
     }
@@ -122,7 +123,7 @@ class AuctionForListSearchByTitleAndSortByClosingTimeTest extends AuctionForList
         // when
         final Slice<Auction> actual = querydslAuctionRepository.findAuctionsAllByCondition(
                 검색어_맥북,
-                PageRequest.of(5, 페이지_크기_3, 마감_임박순_정렬)
+                PageRequest.of(5, 페이지_크기_3, id순_정렬)
         );
 
         // then
